backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
groq==0.4.1
pymongo==4.6.0
certifi==2023.11.17
python-dotenv==1.0.0
